refactored components & styles to make them more modular, and removed useless stuff
