+++ components/Footer/Footer.module.scss
@@ -1,0 +1,6 @@
+$blue: #1f92a2;
+
+.footer {
+  background-color: $blue;
+  padding: 1.4rem;
+}
