+++ components/Navigation/Navigation.module.scss
@@ -1,0 +1,37 @@
+$blue: #1f92a2;
+$purple: #8980f5;
+$white: #edf3f5;
+
+// Group elements that share the same base background
+.navbar,
+.navbarMenu,
+.navbarItem {
+  background-color: $blue;
+}
+
+.navbarItem {
+  transition: background-color 0.3s ease;
+
+  &:hover {
+    // !important is likely needed to override Bulma's global .navbar-item:hover
+    background-color: $purple !important;
+    color: white;
+  }
+}
+
+.navbarBurger {
+  // This targets the three lines (spans) inside the burger
+  span {
+    background-color: white !important;
+  }
+
+  // Change color when the burger is hovered or active
+  &:hover,
+  &.is-active {
+    background-color: $purple !important;
+  }
+}
+
+.logo {
+  margin: 10px;
+}
