+++ components/PageSections/About/About.module.scss
@@ -1,0 +1,20 @@
+$blue: #1f92a2;
+
+.about {
+  border: 4px solid $blue;
+  border-radius: 10px;
+
+  @media (max-width: 767px) {
+    display: flex;
+    flex-direction: column-reverse;
+  }
+
+  .logo img {
+    background-color: white;
+    padding: 2px;
+    width: 50%;
+    margin: 0 auto;
+    display: block;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2), 0 6px 20px rgba(0, 0, 0, 0.19);
+  }
+}
