+++ components/PageSections/Experience/Experience.module.scss
@@ -1,0 +1,17 @@
+$blue: #1f92a2;
+
+.education {
+  border: 4px solid $blue;
+  border-radius: 10px;
+
+  .logo {
+    img {
+      background-color: white;
+      padding: 2px;
+      width: 50%;
+      margin: 0 auto;
+      box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2),
+      0 6px 20px 0 rgba(0, 0, 0, 0.19);
+    }
+  }
+}
