+++ components/PageSections/Header/Header.module.scss
@@ -1,0 +1,23 @@
+$blue: #1f92a2;
+
+.intro {
+  border: 4px solid $blue;
+  border-radius: 10px;
+
+  .name {
+    font-size: 36px;
+  }
+
+  .headshot {
+    width: 50%;
+    margin: 0 auto;
+
+    img {
+      background-color: white;
+      padding: 2px;
+      margin: 0 auto;
+      box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2),
+      0 6px 20px 0 rgba(0, 0, 0, 0.19);
+    }
+  }
+}
