+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import "../styles/globals.scss";
+import "../styles/custom.scss";
 import "aos/dist/aos.css";
 import type { AppProps } from "next/app";
 import { Analytics } from "@vercel/analytics/react";
