+++ pages/index.tsx
@@ -2,18 +2,18 @@ import type { NextPage } from "next";
 import Navigation from "../components/Navigation";
 import Metadata from "../components/Metadata";
 import Footer from "../components/Footer";
-import AboutSection from "../components/sections/About";
-import ProjectSection from "../components/sections/projects/Projects";
-import HeaderSection from "../components/sections/Header";
-import Education from "../components/sections/Experience";
+import About from "../components/PageSections/About";
+import ProjectSection from "../components/PageSections/projects/Projects";
+import Header from "../components/PageSections/Header";
+import Education from "../components/PageSections/Experience";
 
 const Home: NextPage = () => {
   return (
     <>
       <Metadata />
       <Navigation />
-      <HeaderSection />
-      <AboutSection />
+      <Header />
+      <About />
       <Education />
       <Footer />
     </>
