--- styles/Home.module.scss
@@ -1,53 +1,0 @@
-$gray: #48474a;
-$blue: #1f92a2;
-$green: #a2c1c2;
-$white: #edf3f5;
-$purple: #8980f5;
-
-$sectionBorderSize: 4px;
-
-.intro {
-  border: $sectionBorderSize solid $blue;
-  border-radius: 10px;
-
-  .name {
-    font-size: 36px;
-  }
-
-  .headshot {
-    width: 50%;
-    margin: 0 auto;
-
-    img {
-      background-color: white;
-      padding: 2px;
-      margin: 0 auto;
-      box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2),
-        0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    }
-  }
-}
-
-.about,
-.education {
-  border: $sectionBorderSize solid $blue;
-  border-radius: 10px;
-
-  .logo {
-    img {
-      background-color: white;
-      padding: 2px;
-      width: 50%;
-      margin: 0 auto;
-      box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2),
-        0 6px 20px 0 rgba(0, 0, 0, 0.19);
-    }
-  }
-}
-
-.about {
-  @media (max-width: 767px) {
-    display: flex;
-    flex-direction: column-reverse;
-  }
-}
