--- styles/components/Navigation.module.scss
@@ -1,3 +1,0 @@
-.logo {
-  margin: 10px;
-}
