--- styles/globals.scss
@@ -1,46 +1,0 @@
-@import "custom.scss";
-
-html,
-body {
-  padding: 0;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-  color: $gray;
-  background-color: $green;
-  scroll-behavior: smooth;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
-
-a:hover {
-  color: $purple;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-.icon {
-  margin: 0 1rem;
-}
-
-svg.icon {
-  path {
-    fill: $purple;
-  }
-}
-
-/*@media (prefers-color-scheme: dark) {
-  html {
-    color-scheme: dark;
-  }
-  body {
-    color: white;
-    background: black;
-  }
-}
-*/
